requirements.txt
flask==3.0.1
pyodbc==5.0.1
python-dotenv==1.0.0
